requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
opencv-python-headless
numpy
